Modified EntityTexturesConverter converter version
