+++ src/main/java/net/mcreator/element/GeneratableElement.java
@@ -40,7 +40,7 @@ public abstract class GeneratableElement {
 
 	private transient ModElement element;
 
-	public static final transient int formatVersion = 15;
+	public static final transient int formatVersion = 16;
 
 	public GeneratableElement(ModElement element) {
 		if (element != null)
