+++ src/main/java/net/mcreator/element/converter/ConverterRegistry.java
@@ -27,7 +27,7 @@ import net.mcreator.element.converter.fv13.ProcedureSpawnGemPickupDelayFixer;
 import net.mcreator.element.converter.fv14.BlockLuminanceFixer;
 import net.mcreator.element.converter.fv14.DimensionLuminanceFixer;
 import net.mcreator.element.converter.fv14.PlantLuminanceFixer;
-import net.mcreator.element.converter.fv15.EntityTexturesConverter;
+import net.mcreator.element.converter.fv16.EntityTexturesConverter;
 import net.mcreator.element.converter.fv4.RecipeTypeConverter;
 import net.mcreator.element.converter.fv5.AchievementFixer;
 import net.mcreator.element.converter.fv6.GUIBindingInverter;
